smarthome 0.3.0: 1. added aggregator service
