+++ telemetry/aggregator/src/main/java/ru/yandex/practicum/AggregatorApplication.java
@@ -1,0 +1,20 @@
+package ru.yandex.practicum;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+import org.springframework.context.ConfigurableApplicationContext;
+import ru.yandex.practicum.starter.AggregatorStarter;
+
+@SpringBootApplication
+@ConfigurationPropertiesScan
+
+public class AggregatorApplication {
+
+    public static void main(String[] args) {
+        ConfigurableApplicationContext context = SpringApplication.run(AggregatorApplication.class, args);
+        AggregatorStarter aggregator = context.getBean(AggregatorStarter.class);
+        aggregator.start();
+    }
+
+}
