+++ telemetry/aggregator/src/main/java/ru/yandex/practicum/serializer/AggregatorSerializer.java
@@ -1,0 +1,37 @@
+package ru.yandex.practicum.serializer;
+
+import org.apache.avro.io.BinaryEncoder;
+import org.apache.avro.io.DatumWriter;
+import org.apache.avro.io.EncoderFactory;
+import org.apache.avro.specific.SpecificDatumWriter;
+import org.apache.avro.specific.SpecificRecordBase;
+import org.apache.kafka.common.errors.SerializationException;
+import org.apache.kafka.common.serialization.Serializer;
+
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+
+public class AggregatorSerializer implements Serializer<SpecificRecordBase> {
+
+    private final EncoderFactory encoderFactory = EncoderFactory.get();
+    private BinaryEncoder encoder;
+
+
+    @Override
+    public byte[] serialize(String topic, SpecificRecordBase data) {
+        try (ByteArrayOutputStream out = new ByteArrayOutputStream()) {
+            byte[] result = null;
+            encoder = encoderFactory.binaryEncoder(out, encoder);
+            if (data != null) {
+                DatumWriter<SpecificRecordBase> writer = new SpecificDatumWriter<>(data.getSchema());
+                writer.write(data, encoder);
+                encoder.flush();
+                result = out.toByteArray();
+            }
+            return result;
+        } catch (IOException ex) {
+            throw new SerializationException("Ошибка сериализации данных для топика [" + topic + "]", ex);
+        }
+    }
+
+}
