+++ telemetry/aggregator/src/main/java/ru/yandex/practicum/serializer/AggregatorTopics.java
@@ -1,0 +1,7 @@
+package ru.yandex.practicum.serializer;
+
+public class AggregatorTopics {
+
+    public static final String TELEMETRY_SNAPSHOTS_V1 = "telemetry.snapshots.v1";
+    public static final String TELEMETRY_SENSORS_V1 = "telemetry.sensors.v1";
+}
