--- telemetry/collector/src/main/java/ru/practicum/smarthome/CollectorController.java
@@ -1,80 +1,0 @@
-package ru.practicum.smarthome;
-
-import com.google.protobuf.Empty;
-import io.grpc.Status;
-import io.grpc.StatusRuntimeException;
-import io.grpc.stub.StreamObserver;
-import net.devh.boot.grpc.server.service.GrpcService;
-import ru.practicum.smarthome.handlers.hubs.HubEventHandler;
-import ru.practicum.smarthome.handlers.sensors.SensorEventHandler;
-import ru.yandex.practicum.grpc.telemetry.collector.CollectorControllerGrpc;
-import ru.yandex.practicum.grpc.telemetry.event.HubEventProto;
-import ru.yandex.practicum.grpc.telemetry.event.SensorEventProto;
-
-import java.util.Map;
-import java.util.Set;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
-@GrpcService
-public class CollectorController extends CollectorControllerGrpc.CollectorControllerImplBase {
-
-    private final Map<SensorEventProto.PayloadCase, SensorEventHandler> sensorEventHandlers;
-    private final Map<HubEventProto.PayloadCase, HubEventHandler> hubEventHandlers;
-
-    public CollectorController(Set<SensorEventHandler> sensorEventHandlers, Set<HubEventHandler> hubEventHandlers) {
-        this.sensorEventHandlers = sensorEventHandlers.stream()
-                .collect(Collectors.toMap(
-                        SensorEventHandler::getMessageType,
-                        Function.identity()
-                ));
-        this.hubEventHandlers = hubEventHandlers.stream()
-                .collect(Collectors.toMap(
-                        HubEventHandler::getMessageType,
-                        Function.identity()
-                ));
-    }
-
-
-    /**
-     * Метод обработки событий от датчиков.
-     * Вызывается при получении нового события от gRPC-клиента.
-     *
-     * @param request          Событие от датчика
-     * @param responseObserver Ответ для клиента
-     */
-    @Override
-    public void collectSensorEvent(SensorEventProto request, StreamObserver<Empty> responseObserver) {
-        try {
-            if (sensorEventHandlers.containsKey(request.getPayloadCase())) {
-                sensorEventHandlers.get(request.getPayloadCase()).handle(request);
-            } else {
-                throw new IllegalArgumentException("Не могу найти обработчик для события датчика" + request.getPayloadCase());
-            }
-            responseObserver.onNext(Empty.getDefaultInstance());
-            responseObserver.onCompleted();
-        } catch (Exception e) {
-            responseObserver.onError(new StatusRuntimeException(Status.fromThrowable(e)));
-        }
-    }
-
-    @Override
-    public void collectHubEvent(HubEventProto request, StreamObserver<Empty> responseObserver) {
-        try {
-            if (hubEventHandlers.containsKey(request.getPayloadCase())) {
-                hubEventHandlers.get(request.getPayloadCase()).handle(request);
-            } else {
-                throw new IllegalArgumentException("Не могу найти обработчик для события хаба" + request.getPayloadCase());
-            }
-            responseObserver.onNext(Empty.getDefaultInstance());
-            responseObserver.onCompleted();
-        } catch (Exception e) {
-            responseObserver.onError(new StatusRuntimeException(
-                    Status.INTERNAL
-                            .withDescription(e.getLocalizedMessage())
-                            .withCause(e)
-            ));
-        }
-    }
-
-}
