--- telemetry/collector/src/main/java/ru/practicum/smarthome/TelemetryTopics.java
@@ -1,8 +1,0 @@
-package ru.practicum.smarthome;
-
-public class TelemetryTopics {
-
-    public static final String TELEMETRY_SENSORS_V1 = "telemetry.sensors.v1";
-    public static final String TELEMETRY_HUBS_V1 = "telemetry.hubs.v1";
-
-}
