--- telemetry/collector/src/main/java/ru/practicum/smarthome/handlers/sensors/MotionSensorHandler.java
@@ -1,40 +1,0 @@
-package ru.practicum.smarthome.handlers.sensors;
-
-import org.springframework.stereotype.Component;
-import ru.practicum.smarthome.mapper.SensorAvroMapper;
-import ru.yandex.practicum.grpc.telemetry.event.MotionSensorProto;
-import ru.yandex.practicum.grpc.telemetry.event.SensorEventProto;
-import ru.yandex.practicum.kafka.telemetry.event.MotionSensorAvro;
-import ru.yandex.practicum.kafka.telemetry.event.SensorEventAvro;
-
-import java.time.Instant;
-
-import static ru.practicum.smarthome.TelemetryTopics.TELEMETRY_SENSORS_V1;
-
-@Component
-public class MotionSensorHandler extends SensorEventHandler {
-
-    public MotionSensorHandler(SensorAvroMapper sensorAvroMapper) {
-        super(sensorAvroMapper);
-    }
-
-    @Override
-    public SensorEventProto.PayloadCase getMessageType() {
-        return SensorEventProto.PayloadCase.MOTION_SENSOR_EVENT;
-    }
-
-    @Override
-    public void handle(SensorEventProto event) {
-        MotionSensorProto motionSensorProto = event.getMotionSensorEvent();
-        MotionSensorAvro motionSensorAvro = sensorAvroMapper.motionSensorToAvro(motionSensorProto);
-        SensorEventAvro eventAvro = SensorEventAvro.newBuilder()
-                .setId(event.getId())
-                .setHubId(event.getHubId())
-                .setTimestamp(Instant.ofEpochSecond(event.getTimestampOrBuilder().getSeconds(),
-                        event.getTimestampOrBuilder().getNanos()))
-                .setPayload(motionSensorAvro)
-                .build();
-        sendToKafka(TELEMETRY_SENSORS_V1, getMessageType().name(), eventAvro);
-    }
-
-}
