+++ telemetry/collector/src/main/java/ru/yandex/practicum/serializer/CollectorTopics.java
@@ -1,0 +1,8 @@
+package ru.yandex.practicum.serializer;
+
+public class CollectorTopics {
+
+    public static final String TELEMETRY_SENSORS_V1 = "telemetry.sensors.v1";
+    public static final String TELEMETRY_HUBS_V1 = "telemetry.hubs.v1";
+
+}
